Move NullEnum to be package protected in the org.mapstruct package
