+++ core/src/main/java/org/mapstruct/NullEnum.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.mapstruct.util;
+package org.mapstruct;
 
 /**
  * To be used as a default value for enum class annotation elements.
@@ -11,5 +11,5 @@ package org.mapstruct.util;
  * @author Ben Zegveld
  * @since 1.6
  */
-public enum NullEnum {
+enum NullEnum {
 }
